requirements.txt
## Core
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# AI / LangGraph
langgraph==0.0.40
langchain==0.1.4
langchain-groq==0.0.3
langchain-core==0.1.17

# Data & ML
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.12.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Geospatial
geopy==2.4.1
shapely==2.0.2

# Visualization & Monitoring
matplotlib==3.8.2
plotly==5.18.0

# Auth & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utils
loguru==0.7.2
tenacity==8.2.3
apscheduler==3.10.4
